+++ server/.env.example
@@ -1,4 +1,4 @@
-PORT=xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
-MONGODB_URL=xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
-CLERK_SECRET_KEY=xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
-CLERK_PUBLISHABLE_KEY=xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx 
+PORT=5000
+MONGODB_URI=mongodb://localhost:27017/password-manager
+CLERK_SECRET_KEY=
+CLERK_PUBLISHABLE_KEY=
